EcoQuad/payment-backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
razorpay==2.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
